Fee Calc adjustment for appended data
Previous code was more generous to the miners by giving a bit more in fees. This was causing issues with our backend code that only allows for a 5 satoshi buffer over the exact cost of the transaction.

The change includes adding in a bytes conversion for the data string value to get a more accurate size and costing.