Add in array format for issue and transfer data
These changes will allow for the data field in the issueInfo object to be an array. We will then convert all elements in this array to hex strings but in ASM format that includes the spaces.

Transfer in regards to STAS-789 token we require that the data can also be an array. This requires that it is appended to a hex string. For this we create a standard hex string using the BSV library tools. We take the hex representation of the script to be used to append to the previous token script.

Unlocking script is also updated to take the full hex string of the data elements in the array